requirements.txt
PyPDF2==3.0.1
PyYAML==6.0.2